perf(core): drop named-theme catalog from the headless graph
Headless runPipeline and runScene keep default + void only. The full
@ggsvelte/core and /render entries still inject the catalog, so named
themes need no new caller argument.

Lean canvas scatter gzip 124255 → 122557 bytes (-1.7 KB).
