+++ .changeset/slim-theme-catalog.md
@@ -1,0 +1,5 @@
+---
+"@ggsvelte/core": patch
+---
+
+Headless `runPipeline` now ships only the `default` and `void` theme tables. Named themes still resolve from `@ggsvelte/core` and `@ggsvelte/core/render` with no new caller argument.
