+++ packages/core/src/dom/canvas-marks-points.ts
@@ -4,8 +4,8 @@
  */
 import { resolvePointMark, type PointShapeGeometry } from "../mark-style.js";
 import type { PointsBatch } from "../scene.js";
-import type { ThemeTokens } from "../theme.js";
-import { themeVar } from "../theme.js";
+import type { ThemeTokens } from "../theme-construct.js";
+import { themeVar } from "../theme-resolve.js";
 import type { ColorResolver } from "./canvas-dom.js";
 import { maskIncludes, type PrimitiveFocusMask } from "./canvas-marks-mask.js";
