+++ packages/core/src/dom/canvas-marks-segments.ts
@@ -9,8 +9,8 @@
 import { canvasGradientStyle, subpathBounds, type ResolvedGradientPaint } from "../mark-paint.js";
 import { resolveSegmentMark, segmentStrokeAt } from "../mark-style.js";
 import type { SegmentsBatch } from "../scene.js";
-import type { ThemeTokens } from "../theme.js";
-import { themeVar } from "../theme.js";
+import type { ThemeTokens } from "../theme-construct.js";
+import { themeVar } from "../theme-resolve.js";
 import type { ColorResolver } from "./canvas-dom.js";
 import { applyDash } from "./canvas-marks-paths.js";
