+++ packages/core/src/editions.ts
@@ -21,22 +21,16 @@
  * an `unknown-edition` warning — old renderers degrade to their best defaults
  * instead of failing.
  */
-import type { ThemeName } from "@ggsvelte/spec";
-
-import { VIRIDIS_RAMP_10 } from "./scales/color.js";
+import { VIRIDIS_RAMP_10 } from "./scales/viridis-ramp.js";
 import { CATEGORICAL_PALETTE_10 } from "./scales/train.js";
-import type { ThemeTokens } from "./theme.js";
-import { BUILTIN_THEMES, LEGACY_BUILTIN_THEMES } from "./theme.js";
+import { BUILTIN_THEMES, LEGACY_BUILTIN_THEMES } from "./theme-builtins.js";
+import {
+  resolveEditionDefaults as resolveEditionDefaultsAgainst,
+  type EditionDefaults,
+  type ResolvedEdition,
+} from "./editions-resolve.js";
 
-/** The default-aesthetics bundle one edition pins. */
-export interface EditionDefaults {
-  /** Default range for discrete color/fill scales (no scheme/range set). */
-  categoricalPalette: readonly string[];
-  /** Default ramp stops for sequential color/fill scales. */
-  sequentialRamp: readonly string[];
-  /** Built-in theme tokens the theme registry resolves names against. */
-  themes: Readonly<Record<ThemeName, ThemeTokens>>;
-}
+export type { EditionDefaults, ResolvedEdition } from "./editions-resolve.js";
 
 /** Editions known to this build. Edition 1 = the 0.1.0 defaults. */
 export const EDITION_DEFAULTS: Readonly<Record<number, EditionDefaults>> = Object.freeze({
@@ -52,14 +46,6 @@ export const EDITION_DEFAULTS: Readonly<Record<number, EditionDefaults>> = Objec
   }),
 });
 
-export interface ResolvedEdition {
-  /** The edition whose defaults apply (after fallback). */
-  edition: number;
-  defaults: EditionDefaults;
-  /** Set when the requested edition is unknown to the table (fallback used). */
-  unknownRequested: number | null;
-}
-
 /**
  * Resolve the defaults for a spec's edition. `undefined` means "current"
  * (normalize stamps specs, but non-validating callers may skip it); unknown
@@ -69,16 +55,5 @@ export function resolveEditionDefaults(
   edition?: number,
   table: Readonly<Record<number, EditionDefaults>> = EDITION_DEFAULTS,
 ): ResolvedEdition {
-  const known = Object.keys(table)
-    .map(Number)
-    .filter((n) => Number.isFinite(n))
-    .toSorted((a, b) => a - b);
-  if (known.length === 0) throw new Error("resolveEditionDefaults: empty edition table");
-  const latest = known.at(-1)!;
-  const requested = edition ?? latest;
-  const defaults = table[requested];
-  if (defaults !== undefined) {
-    return { edition: requested, defaults, unknownRequested: null };
-  }
-  return { edition: latest, defaults: table[latest]!, unknownRequested: requested };
+  return resolveEditionDefaultsAgainst(edition, table);
 }
