+++ packages/core/src/index.ts
@@ -342,7 +342,7 @@ export type {
 
 // Pipeline
 export { batchMarkCount, CANVAS_AUTO_THRESHOLD, PipelineError } from "./pipeline/public-api.js";
-export { runPipeline } from "./pipeline/run-pipeline.js";
+export { runPipeline } from "./pipeline/run-pipeline-full.js";
 export type {
   Advisory,
   AxisValueFormatter,
